# CarRent.Api/OpenAPI/Models/Car.cs
/*
 * OpenAPI car rent
 *
 * car rent api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using OpenAPI.Converters;

namespace OpenAPI.Models
{ 
    /// <summary>
    /// A car
    /// </summary>
    [DataContract]
    public class Car : IEquatable<Car>
    {
        /// <summary>
        /// Gets or Sets IdCar
        /// </summary>
        [DataMember(Name="idCar", EmitDefaultValue=false)]
        public long IdCar { get; set; }

        /// <summary>
        /// Gets or Sets CarMake
        /// </summary>
        [DataMember(Name="carMake", EmitDefaultValue=false)]
        public CarMake CarMake { get; set; }

        /// <summary>
        /// Gets or Sets CarName
        /// </summary>
        [DataMember(Name="carName", EmitDefaultValue=false)]
        public string CarName { get; set; }

        /// <summary>
        /// Gets or Sets CarNr
        /// </summary>
        [DataMember(Name="carNr", EmitDefaultValue=false)]
        public string CarNr { get; set; }

        /// <summary>
        /// Gets or Sets CarType
        /// </summary>
        [DataMember(Name="carType", EmitDefaultValue=false)]
        public CarType CarType { get; set; }

        /// <summary>
        /// Gets or Sets CarClass
        /// </summary>
        [DataMember(Name="carClass", EmitDefaultValue=false)]
        public CarClass CarClass { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Car {\n");
            sb.Append("  IdCar: ").Append(IdCar).Append("\n");
            sb.Append("  CarMake: ").Append(CarMake).Append("\n");
            sb.Append("  CarName: ").Append(CarName).Append("\n");
            sb.Append("  CarNr: ").Append(CarNr).Append("\n");
            sb.Append("  CarType: ").Append(CarType).Append("\n");
            sb.Append("  CarClass: ").Append(CarClass).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Car)obj);
        }

        /// <summary>
        /// Returns true if Car instances are equal
        /// </summary>
        /// <param name="other">Instance of Car to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Car other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    IdCar == other.IdCar ||
                    
                    IdCar.Equals(other.IdCar)
                ) && 
                (
                    CarMake == other.CarMake ||
                    CarMake != null &&
                    CarMake.Equals(other.CarMake)
                ) && 
                (
                    CarName == other.CarName ||
                    CarName != null &&
                    CarName.Equals(other.CarName)
                ) && 
                (
                    CarNr == other.CarNr ||
                    CarNr != null &&
                    CarNr.Equals(other.CarNr)
                ) && 
                (
                    CarType == other.CarType ||
                    CarType != null &&
                    CarType.Equals(other.CarType)
                ) && 
                (
                    CarClass == other.CarClass ||
                    CarClass != null &&
                    CarClass.Equals(other.CarClass)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    
                    hashCode = hashCode * 59 + IdCar.GetHashCode();
                    if (CarMake != null)
                    hashCode = hashCode * 59 + CarMake.GetHashCode();
                    if (CarName != null)
                    hashCode = hashCode * 59 + CarName.GetHashCode();
                    if (CarNr != null)
                    hashCode = hashCode * 59 + CarNr.GetHashCode();
                    if (CarType != null)
                    hashCode = hashCode * 59 + CarType.GetHashCode();
                    if (CarClass != null)
                    hashCode = hashCode * 59 + CarClass.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(Car left, Car right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Car left, Car right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}