# CarRent.Api/OpenAPI/Models/CarMake.cs
/*
 * OpenAPI car rent
 *
 * car rent api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using OpenAPI.Converters;

namespace OpenAPI.Models
{ 
    /// <summary>
    /// A car make class
    /// </summary>
    [DataContract]
    public class CarMake : IEquatable<CarMake>
    {
        /// <summary>
        /// Gets or Sets IdCarMake
        /// </summary>
        [DataMember(Name="idCarMake", EmitDefaultValue=false)]
        public long IdCarMake { get; set; }

        /// <summary>
        /// Gets or Sets _CarMake
        /// </summary>
        [DataMember(Name="carMake", EmitDefaultValue=false)]
        public string _CarMake { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class CarMake {\n");
            sb.Append("  IdCarMake: ").Append(IdCarMake).Append("\n");
            sb.Append("  _CarMake: ").Append(_CarMake).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((CarMake)obj);
        }

        /// <summary>
        /// Returns true if CarMake instances are equal
        /// </summary>
        /// <param name="other">Instance of CarMake to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(CarMake other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    IdCarMake == other.IdCarMake ||
                    
                    IdCarMake.Equals(other.IdCarMake)
                ) && 
                (
                    _CarMake == other._CarMake ||
                    _CarMake != null &&
                    _CarMake.Equals(other._CarMake)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    
                    hashCode = hashCode * 59 + IdCarMake.GetHashCode();
                    if (_CarMake != null)
                    hashCode = hashCode * 59 + _CarMake.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(CarMake left, CarMake right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(CarMake left, CarMake right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}